Rename LayoutSpinnerComponent to SpinnerComponent and refactor implementing interface in Scheduling.BlazorClient.Client #5
