--- Scheduling/src/Presentation/BlazorClient/Fnunez.VeterinaryClinic.Scheduling.BlazorClient.Client/Shared/Components/LayoutSpinner/LayoutSpinner.razor
@@ -1,8 +1,0 @@
-@inherits Fnunez.VeterinaryClinic.Scheduling.BlazorClient.Client.Shared.Components.LayoutSpinnerComponent;
-
-@if (IsVisible)
-{
-    <div class="spinner-overlay">
-        <div class="spinner-border"></div>
-    </div>
-}
