+++ Scheduling/src/Presentation/BlazorClient/Fnunez.VeterinaryClinic.Scheduling.BlazorClient.Client/Shared/Components/Spinner/SpinnerComponent.razor
@@ -1,0 +1,6 @@
+@if (IsVisible)
+{
+    <div class="spinner-overlay">
+        <div class="spinner-border"></div>
+    </div>
+}
