+++ Scheduling/src/Presentation/BlazorClient/Fnunez.VeterinaryClinic.Scheduling.BlazorClient.Client/Shared/Components/Spinner/SpinnerComponent.razor.cs
@@ -1,0 +1,5 @@
+namespace Fnunez.VeterinaryClinic.Scheduling.BlazorClient.Client.Shared.Components.Spinner;
+
+public partial class SpinnerComponent
+{
+}
